monitor: Rename files to match their content
This commit exists to improve the readability of the previous commit,
which renamed the Monitor struct to Agent and the Agent struct to
server.
